remove unused/moved code from SDK
