--- src/aleph_client/vm/__init__.py
@@ -1,3 +1,0 @@
-"""
-Aleph helpers for apps running inside aleph.im Virtual Machines.
-"""
