--- tests/integration/config.py
@@ -1,3 +1,0 @@
-TARGET_NODE = "http://163.172.70.92:4024"
-REFERENCE_NODE = "https://api2.aleph.im"
-TEST_CHANNEL = "INTEGRATION_TESTS"
