--- tests/integration/pytest.ini
@@ -1,2 +1,0 @@
-[pytest]
-python_files = itest_*.py
