+++ tests/unit/conftest.py
@@ -8,36 +8,20 @@
 """
 from pathlib import Path
 from tempfile import NamedTemporaryFile
+from typing import Generator
 
-import aleph.sdk.chains.ethereum as ethereum
-import aleph.sdk.chains.sol as solana
-import aleph.sdk.chains.tezos as tezos
 import pytest
-from aleph.sdk.chains.common import get_fallback_private_key
+from aleph.sdk.chains.common import generate_key
 
 
 @pytest.fixture
-def fallback_private_key() -> bytes:
-    with NamedTemporaryFile() as private_key_file:
-        yield get_fallback_private_key(path=Path(private_key_file.name))
+def empty_account_file() -> Generator[Path, None, None]:
+    with NamedTemporaryFile() as key_file:
+        yield Path(key_file.name)
 
 
 @pytest.fixture
-def ethereum_account() -> ethereum.ETHAccount:
-    with NamedTemporaryFile(delete=False) as private_key_file:
-        private_key_file.close()
-        yield ethereum.get_fallback_account(path=Path(private_key_file.name))
-
-
-@pytest.fixture
-def solana_account() -> solana.SOLAccount:
-    with NamedTemporaryFile(delete=False) as private_key_file:
-        private_key_file.close()
-        yield solana.get_fallback_account(path=Path(private_key_file.name))
-
-
-@pytest.fixture
-def tezos_account() -> tezos.TezosAccount:
-    with NamedTemporaryFile(delete=False) as private_key_file:
-        private_key_file.close()
-        yield tezos.get_fallback_account(path=Path(private_key_file.name))
+def account_file(empty_account_file: Path) -> Path:
+    private_key = generate_key()
+    empty_account_file.write_bytes(private_key)
+    return empty_account_file
